refactor UserFilter
