+++ src/main/java/school/faang/user_service/controller/SubscriptionController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.dto.UserFilterDto;
-import school.faang.user_service.entity.UserFilter;
+import school.faang.user_service.filter.UserFilter;
 import school.faang.user_service.service.SubscriptionService;
 
 import java.util.List;
